Refactor feedback question response attributes
